correctly translate docker image pull statuses instead of using them directly
