+++ applications/launchpad/gui-react/src/containers/SettingsContainer/DockerSettings/DockerImagesList/styles.ts
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 
 export const DockerRow = styled.div`
   display: flex;
-  padding: ${({ theme }) => theme.spacingVertical()};
+  padding: ${({ theme }) => theme.spacingVertical(1.25)};
   &:not(:last-of-type) {
     border-bottom: 1px solid ${({ theme }) => theme.borderColor};
   }
